refactor(logging, sync): Implement custom logger and refine WebSocket sync logic
This commit introduces a centralized `AppLog` utility to wrap Android logging with debug-gated checks. It also updates various repositories and the synchronization orchestrator to utilize the new logging utility, removes legacy sync methods, and fixes several issues related to document versioning and full sync behavior.

### Key Changes:

-   **Core Utilities:**
    -   **`AppLog.kt`**: Introduced a new logging wrapper that performs no-ops in release builds, ensuring sensitive debug information is not leaked.
    -   **`Constants.kt`**: Updated `BASE_URL` to use `BuildConfig.BACK_BASE_URL` for better build-type configuration.

-   **Database & DI:**
    -   **`DatabaseModule.kt`**: Removed hardcoded demo data population logic. The database now relies on server synchronization or specialized demo builds rather than static inserts during creation.

-   **Sync & WebSocket (`SyncOrchestrator.kt`, `WebSocketManager.kt`):**
    -   **Document Versioning**: Updated `version` fields from `Int` to `Long` across `SyncMessage` and `SyncOrchestrator` to prevent overflow issues and maintain consistency with server-side data types.
    -   **Full Sync Logic**: In `SettingsViewModel`, cache clearing now explicitly deletes sync state and triggers a full WebSocket synchronization request.
    -   **Pending Operations**: Improved the robustness of `processPendingOperations` by adding better state checks and error handling for offline-to-online transitions.
    -   **Cleanup**: Removed deprecated legacy sync methods (`syncAll`, `syncEntity`) from repositories and the orchestrator.

-   **Repositories:**
    -   **`ClientRepository`**, **`ProductRepository`**, **`WarehouseRepository`**: Removed `syncX()` interface methods as synchronization is now centrally managed via the `SyncOrchestrator` over WebSocket.
    -   **`UserRepositoryImpl`**: Refined the login flow to better handle the transition between WebSocket authentication and offline fallback, using the new `AppLog` for traceability.

-   **Barcode & Scanning:**
    -   **`BarcodeService.kt`**, **`HardwareScannerManager.kt`**, **`CameraScannerManager.kt`**, **`GS1Parser.kt`**: Migrated all logging from `android.util.Log` to `AppLog`.

-   **Build Configuration:**
    -   Updated Gradle wrapper to `9.3.1` and Android Gradle Plugin (AGP) to `9.1.0`.

+++ app/src/main/java/ua/com/programmer/pick/data/remote/websocket/SyncMessage.kt
@@ -255,7 +255,7 @@ sealed class SyncMessage {
         val success: Boolean,
         val state: String? = null,
         val completedAt: String? = null,
-        val version: Int? = null,
+        val version: Long? = null,
         val error: String? = null
     ) : SyncMessage() {
         override val type = MessageType.DOCUMENT_COMPLETE_RESULT
